Detox on Android (#4098)
