+++ .detoxrc.js
@@ -39,7 +39,14 @@ module.exports = {
       build:
         'cd ./android && ./gradlew assembleDebug assembleAndroidTest -DtestBuildType=debug && cd ..',
       type: 'android.emulator',
-      name: 'Pixel_3_API_29',
+      name: 'Pixel_5_API_31',
+    },
+    'android.emu.release': {
+      type: 'android.emulator',
+      binaryPath: 'android/app/build/outputs/apk/release/app-release.apk',
+      build:
+        'cd android && ./gradlew assembleRelease assembleAndroidTest -DtestBuildType=release && cd ..',
+      name: 'Pixel_5_API_31',
     },
   },
 };
