+++ .github/workflows/macstadium-android-e2e.yml
@@ -9,7 +9,7 @@ on: [pull_request, workflow_dispatch]
 jobs:
   # This workflow contains a single job called "android-e2e"
   android-e2e:
-    if: github.ref_name == '@tchayen/android-detox'
+    if: github.head_ref == '@tchayen/android-detox'
     # The type of runner that the job will run on
     runs-on: ["self-hosted", "ci-5"]
     # Cancel current builds if there's a newer commit on the same branch
@@ -49,7 +49,7 @@ jobs:
         run: ./node_modules/.bin/detox build --configuration android.emu.release
 
       - name: Run Android e2e tests
-        run: ./node_modules/.bin/detox test -R 5 --configuration android.emu.release --forceExit
+        run: ./node_modules/.bin/detox test -R 5 -w 1 --configuration android.emu.release --forceExit
 
       - name: Send success message to slack
         id: slack-success-msg
