+++ e2e/sendSheetFlow.spec.js
@@ -1,11 +1,11 @@
 /* eslint-disable no-undef */
 /* eslint-disable jest/expect-expect */
-import { exec } from 'child_process';
 import * as Helpers from './helpers';
 
+const android = device.getPlatform() === 'android';
+
 beforeAll(async () => {
-  // Connect to hardhat
-  await exec('yarn hardhat');
+  await Helpers.startHardhat();
 });
 
 describe('Send Sheet Interaction Flow', () => {
@@ -23,16 +23,6 @@ describe('Send Sheet Interaction Flow', () => {
     await Helpers.checkIfExists('import-sheet');
   });
 
-  it("Shouldn't do anything when I type jibberish", async () => {
-    await Helpers.waitAndTap('import-sheet-input');
-    await Helpers.checkIfElementHasString('import-sheet-button-label', 'Paste');
-    await Helpers.typeText('import-sheet-input', 'asdajksdlakjsd', false);
-    await Helpers.checkIfElementHasString(
-      'import-sheet-button-label',
-      'Import'
-    );
-  });
-
   it('Should show the "Add wallet modal" after tapping import with a valid seed"', async () => {
     await Helpers.clearField('import-sheet-input');
     await Helpers.typeText('import-sheet-input', process.env.TEST_SEEDS, false);
@@ -47,7 +37,7 @@ describe('Send Sheet Interaction Flow', () => {
   it('Should navigate to the Wallet screen after tapping on "Import Wallet"', async () => {
     await Helpers.disableSynchronization();
     await Helpers.waitAndTap('wallet-info-submit-button');
-    if (device.getPlatform() === 'android') {
+    if (android) {
       await Helpers.checkIfVisible('pin-authentication-screen');
       // Set the pin
       await Helpers.authenticatePin('1234');
@@ -62,16 +52,9 @@ describe('Send Sheet Interaction Flow', () => {
     await Helpers.sendETHtoTestWallet();
   });
 
-  it('Should connect to hardhat', async () => {
-    await Helpers.swipe('wallet-screen', 'right', 'slow');
-    await Helpers.checkIfVisible('profile-screen');
-    await Helpers.waitAndTap('settings-button');
-    await Helpers.checkIfVisible('settings-sheet');
-    await Helpers.waitAndTap('developer-section');
-    await Helpers.checkIfVisible('developer-settings-sheet');
-    await Helpers.waitAndTap('hardhat-section');
+  it('Should show Hardhat Toast after pressing Connect To Hardhat', async () => {
+    await Helpers.waitAndTap('dev-button-hardhat');
     await Helpers.checkIfVisible('testnet-toast-Hardhat');
-    await Helpers.swipe('profile-screen', 'left', 'slow');
   });
 
   // Saving for now in case we want to test iCloud back up sheet
@@ -127,7 +110,7 @@ describe('Send Sheet Interaction Flow', () => {
       '0xF0f21ab2012731542731df194cfF6c77d29cB31A',
       false
     );
-    await Helpers.checkIfVisible('add-contact-button');
+    // await Helpers.checkIfVisible('add-contact-button');
     await Helpers.checkIfVisible('send-asset-list');
   });
 
@@ -139,15 +122,16 @@ describe('Send Sheet Interaction Flow', () => {
       'neverselling.wallet\n',
       false
     );
-    await Helpers.checkIfVisible('add-contact-button');
     await Helpers.checkIfVisible('send-asset-list');
     await Helpers.clearField('send-asset-form-field');
+    await device.disableSynchronization();
     await Helpers.typeText(
       'send-asset-form-field',
       'rainbowwallet.eth\n',
       false
     );
-    await Helpers.checkIfVisible('add-contact-button');
+    await device.enableSynchronization();
+    // await Helpers.checkIfVisible('add-contact-button')
     await Helpers.checkIfVisible('send-asset-list');
   });
 
@@ -185,11 +169,7 @@ describe('Send Sheet Interaction Flow', () => {
     await Helpers.checkIfVisible('selected-asset-field-input');
     await Helpers.waitAndTap('selected-asset-field-input');
     await Helpers.typeText('selected-asset-field-input', '9999', false);
-    if (device.getPlatform() === 'android') {
-      await Helpers.checkIfElementByTextToExist('Insufficient Funds');
-    } else {
-      await Helpers.checkIfElementByTextIsVisible('Insufficient Funds');
-    }
+    await Helpers.checkIfElementByTextIsVisible('Insufficient Funds');
   });
 
   it('Should prepend a 0 to quantity field on input of .', async () => {
@@ -262,71 +242,9 @@ describe('Send Sheet Interaction Flow', () => {
     await Helpers.checkIfElementByTextIsVisible('8.12');
     await Helpers.waitAndTap('send-asset-form-ETH-token');
   });
-
-  it('Should show Add Contact Screen after tapping Add Contact Button', async () => {
-    await Helpers.checkIfVisible('add-contact-button');
-    await Helpers.waitAndTap('add-contact-button');
-    await Helpers.checkIfVisible('wallet-info-input');
-  });
-
-  it('Should do nothing on Add Contact cancel', async () => {
-    await Helpers.tapByText('Cancel');
-    await Helpers.checkIfVisible('add-contact-button');
-    await Helpers.waitAndTap('add-contact-button');
-    await Helpers.tapByText('Cancel');
-  });
-
-  it('Should update address field to show contact name & show edit contact button', async () => {
-    await Helpers.waitAndTap('add-contact-button');
-    await Helpers.clearField('wallet-info-input');
-    await Helpers.typeText('wallet-info-input', 'testcoin.test', true);
-    await Helpers.waitAndTap('wallet-info-submit-button');
-    await Helpers.checkIfElementByTextIsVisible('testcoin.test');
-    await Helpers.checkIfVisible('edit-contact-button');
-  });
-
-  it('Should show Asset List & Edit Contact Button on cancel', async () => {
-    await Helpers.checkIfVisible('edit-contact-button');
-    await Helpers.waitAndTap('edit-contact-button');
-    await Helpers.tapByText('Cancel');
-  });
-
-  it('Should updated contact name after edit contact', async () => {
-    await Helpers.checkIfVisible('edit-contact-button');
-    await Helpers.waitAndTap('edit-contact-button');
-    await Helpers.tapByText('Edit Contact');
-    await Helpers.clearField('wallet-info-input');
-    await Helpers.typeText('wallet-info-input', 'testcoin.eth', true);
-    await Helpers.waitAndTap('wallet-info-submit-button');
-    // await Helpers.tapByText('Done');
-    await Helpers.checkIfElementByTextIsVisible('testcoin.eth');
-  });
-
-  it('Should load contacts if contacts exist', async () => {
-    if (device.getPlatform() === 'android') {
-      await device.pressBack();
-    } else {
-      await Helpers.swipe('send-asset-form-field', 'down', 'slow');
-    }
-    await Helpers.waitAndTap('send-fab');
-    await Helpers.checkIfElementByTextIsVisible('testcoin.eth');
-  });
-
-  it('Should show Add Contact Button after deleting contact', async () => {
-    await Helpers.checkIfElementByTextIsVisible('testcoin.eth');
-    await Helpers.tapByText('testcoin.eth');
-    await Helpers.checkIfVisible('edit-contact-button');
-    await Helpers.waitAndTap('edit-contact-button');
-    await Helpers.tapByText('Delete Contact');
-    await Helpers.delay(2000);
-    await Helpers.tapByText('Delete Contact');
-    await Helpers.delay(2000);
-    await Helpers.checkIfVisible('add-contact-button');
-  });
-
   afterAll(async () => {
     // Reset the app state
     await device.clearKeychain();
-    await exec('kill $(lsof -t -i:8545)');
+    await Helpers.killHardhat();
   });
 });
