+++ e2e/watchAddressFlow.spec.js
@@ -2,6 +2,8 @@
 /* eslint-disable jest/expect-expect */
 import * as Helpers from './helpers';
 
+const android = device.getPlatform() === 'android';
+
 describe('Watch address flow', () => {
   it('Should show the welcome screen', async () => {
     await Helpers.checkIfVisible('welcome-screen');
@@ -31,19 +33,13 @@ describe('Watch address flow', () => {
 
   it('Should navigate to the Wallet screen after tapping on "Import Wallet"', async () => {
     await Helpers.waitAndTap('wallet-info-submit-button');
-    if (device.getPlatform() === 'android') {
-      await Helpers.checkIfVisible('pin-authentication-screen');
-      // Set the pin
-      await Helpers.authenticatePin('1234');
-      // Confirm it
-      await Helpers.authenticatePin('1234');
-    }
     await Helpers.checkIfVisible('wallet-screen', 40000);
     await Helpers.enableSynchronization();
   });
 
   it('Should say "test.eth" in the Profile Screen header', async () => {
-    if (device.getPlatform() === 'android') {
+    if (android) {
+      // TODO: now is the time to clean this up.
       // not working on android!
       // (app isn't idle and test times out)
       //   // await Helpers.checkIfElementByTextIsVisible('test.eth');
