+++ src/components/animations/ShimmerAnimation.js
@@ -9,6 +9,7 @@ import Animated, {
 } from 'react-native-reanimated';
 import styled from '@/styled-thing';
 import { position } from '@/styles';
+import { IS_TESTING } from 'react-native-dotenv';
 
 const timingConfig = {
   duration: 2500,
@@ -54,6 +55,10 @@ export default function ShimmerAnimation({
     opacity: opacity.value,
     transform: [{ translateX: positionX.value }],
   }));
+
+  if (IS_TESTING === 'true') {
+    return null;
+  }
 
   return <ColorGradient colors={gradientColors} style={animatedStyle} />;
 }
