+++ src/components/discover-sheet/DiscoverSheet.android.js
@@ -132,6 +132,7 @@ const DiscoverSheet = (_, forwardedRef) => {
         >
           <DiscoverSheetHeader yPosition={yPosition} />
           <BottomSheetScrollView
+            testID="discover-sheet"
             onScrollWorklet={scrollHandler}
             ref={sheet}
             removeClippedSubviews
