+++ src/screens/ExchangeModal.tsx
@@ -84,6 +84,7 @@ import Routes from '@/navigation/routesNames';
 import { ethereumUtils, gasUtils } from '@/utils';
 import { useEthUSDPrice } from '@/utils/ethereumUtils';
 import logger from 'logger';
+import { IS_ANDROID, IS_TEST } from '@/env';
 
 export const DEFAULT_SLIPPAGE_BIPS = {
   [Network.mainnet]: 100,
@@ -544,7 +545,11 @@ export default function ExchangeModal({
   }, [updateMaxInputAmount]);
 
   const checkGasVsOutput = async (gasPrice: string, outputPrice: string) => {
-    if (greaterThan(outputPrice, 0) && greaterThan(gasPrice, outputPrice)) {
+    if (
+      greaterThan(outputPrice, 0) &&
+      greaterThan(gasPrice, outputPrice) &&
+      !(IS_ANDROID && IS_TEST)
+    ) {
       const res = new Promise(resolve => {
         Alert.alert(
           lang.t('swap.warning.cost.are_you_sure_title'),
